Mise à jour des configurations Codemagic et pubspec

+++ lib/screens/onboarding/onboarding_screen.dart
@@ -131,7 +131,7 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
                 children: [
                   DotsIndicator(
                     dotsCount: pages.length,
-                    position: _currentPage, // <-- CORRECTION APPLIQUÉE ICI
+                    position: _currentPage.toDouble(), // <-- CORRECTION APPLIQUÉE ICI
                     decorator: DotsDecorator(
                       size: const Size.square(9.0),
                       activeSize: const Size(18.0, 9.0),
